feat(run): Flag to jump to the first failed commit
Fixes #264
